Refector repository API by introducing abstact iterator and removing specialized methods on top level
